cdnmanager/requirements.txt
Django==1.3
South==0.7.3
PIL==1.1.7
django-extensions==0.6
simplejson==2.1.1
pyftpdlib
psycopg2
ftputil
celery
django-celery
paramiko
django-piston==0.2.2
